requirements.txt
setuptools==65.5.0
cloudpickle==1.6.0
cycler==0.12.1
-e git+https://github.com/WE-Autopilot/f1tenth_gym.git@bdee71de114e160ff920563f3b443e10db20e94c#egg=f110_gym&subdirectory=gym
fonttools==4.56.0
gym==0.19.0
kiwisolver==1.4.8
llvmlite==0.43.0
matplotlib==3.5.3
numba==0.60.0
numpy==1.22.0
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pyglet==1.5.20
PyOpenGL==3.1.9
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.11.4
six==1.17.0
torch==2.6.0
-e git+https://github.com/WE-Autopilot/f1tenth_gym.git@bdee71de114e160ff920563f3b443e10db20e94c#egg=weap_util&subdirectory=weap_util